+++ packages/config/src/xcm-configs/moonriver.ts
@@ -13,7 +13,6 @@ import {
   kbtc,
   kint,
   ksm,
-  mgx,
   movr,
   rmrk,
   sdn,
@@ -33,7 +32,6 @@ import {
   kintsugi,
   kusama,
   kusamaAssetHub,
-  mangataKusama,
   moonriver,
   robonomics,
   shiden,
@@ -113,29 +111,6 @@ export const moonriverRoutes = new ChainRoutes({
           asset: movr,
         },
         min: AssetMinBuilder().assetRegistry().assetMetadatas(),
-      },
-      contract: ContractBuilder().XcmPrecompile().transferAssetsToPara32(),
-    },
-    {
-      source: {
-        asset: movr,
-        balance: BalanceBuilder().substrate().system().account(),
-        fee: {
-          asset: movr,
-          balance: BalanceBuilder().substrate().system().account(),
-        },
-        destinationFee: {
-          balance: BalanceBuilder().substrate().system().account(),
-        },
-      },
-      destination: {
-        asset: movr,
-        chain: mangataKusama,
-        balance: BalanceBuilder().substrate().tokens().accounts(),
-        fee: {
-          amount: 0.002,
-          asset: movr,
-        },
       },
       contract: ContractBuilder().XcmPrecompile().transferAssetsToPara32(),
     },
@@ -356,29 +331,6 @@ export const moonriverRoutes = new ChainRoutes({
     },
     {
       source: {
-        asset: mgx,
-        balance: BalanceBuilder().evm().erc20(),
-        fee: {
-          asset: movr,
-          balance: BalanceBuilder().substrate().system().account(),
-        },
-        destinationFee: {
-          balance: BalanceBuilder().evm().erc20(),
-        },
-      },
-      destination: {
-        asset: mgx,
-        chain: mangataKusama,
-        balance: BalanceBuilder().substrate().tokens().accounts(),
-        fee: {
-          amount: 5.5,
-          asset: mgx,
-        },
-      },
-      contract: ContractBuilder().XcmPrecompile().transferAssetsToPara32(),
-    },
-    {
-      source: {
         asset: rmrk,
         balance: BalanceBuilder().evm().erc20(),
         fee: {
@@ -522,7 +474,7 @@ export const moonriverRoutes = new ChainRoutes({
         balance: BalanceBuilder().substrate().tokens().accounts(),
         fee: {
           amount: FeeBuilder().xcmPaymentApi().fromCurrencyIdToLocations({
-            isAssetReserveChain: false,
+            isAssetReserveChain: true,
           }),
           asset: vbnc,
         },
@@ -548,7 +500,7 @@ export const moonriverRoutes = new ChainRoutes({
         balance: BalanceBuilder().substrate().tokens().accounts(),
         fee: {
           amount: FeeBuilder().xcmPaymentApi().fromCurrencyIdToLocations({
-            isAssetReserveChain: false,
+            isAssetReserveChain: true,
           }),
           asset: vksm,
         },
@@ -574,7 +526,7 @@ export const moonriverRoutes = new ChainRoutes({
         balance: BalanceBuilder().substrate().tokens().accounts(),
         fee: {
           amount: FeeBuilder().xcmPaymentApi().fromCurrencyIdToLocations({
-            isAssetReserveChain: false,
+            isAssetReserveChain: true,
           }),
           asset: vmovr,
         },