+++ packages/builder/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @moonbeam-network/xcm-builder
 
+## 3.3.7
+
+### Patch Changes
+
+- [#527](https://github.com/moonbeam-foundation/xcm-sdk/pull/527) [`ead3303`](https://github.com/moonbeam-foundation/xcm-sdk/commit/ead33037b56da906b2db3e4e337af56e64f0cddd) Thanks [@Rihyx](https://github.com/Rihyx)! - Update dependencies
+
+- Updated dependencies [[`ead3303`](https://github.com/moonbeam-foundation/xcm-sdk/commit/ead33037b56da906b2db3e4e337af56e64f0cddd)]:
+  - @moonbeam-network/xcm-types@3.2.7
+
+## 3.3.6
+
+### Patch Changes
+
+- [#520](https://github.com/moonbeam-foundation/xcm-sdk/pull/520) [`8179c3f`](https://github.com/moonbeam-foundation/xcm-sdk/commit/8179c3f97be3e5d9c23db1cce1d868ea013f15e3) Thanks [@mmaurello](https://github.com/mmaurello)! - Change some routes from using xTokens pallet to polkadotXcm
+
+- Updated dependencies [[`8179c3f`](https://github.com/moonbeam-foundation/xcm-sdk/commit/8179c3f97be3e5d9c23db1cce1d868ea013f15e3)]:
+  - @moonbeam-network/xcm-types@3.2.6
+
 ## 3.3.5
 
 ### Patch Changes